Reorganize summarizerNode files to minimize implementation leak
