+++ packages/runtime/runtime-utils/src/index.ts
@@ -10,8 +10,7 @@ export * from "./objectstoragepartition";
 export * from "./objectstorageutils";
 export * from "./requestParser";
 export * from "./serializer";
-export * from "./summarizerNode";
-export * from "./summarizerNodeWithGc";
+export * from "./summarizerNodeContract";
 export * from "./summaryTracker";
 export * from "./summaryUtils";
 export * from "./utils";
